--- a/crates/next-api/src/app.rs
+++ b/crates/next-api/src/app.rs
@@ -1642,9 +1642,15 @@ impl Endpoint for AppEndpoint {
                 .clone_value();
 
             let client_relative_root = this.app_project.project().client_relative_path();
-            let client_paths = all_paths_in_root(output_assets, client_relative_root)
-                .await?
-                .clone_value();
+            let client_paths = async {
+                anyhow::Ok(
+                    all_paths_in_root(output_assets, client_relative_root)
+                        .await?
+                        .clone_value(),
+                )
+            }
+            .instrument(tracing::info_span!("client_paths"))
+            .await?;
 
             let written_endpoint = match *output {
                 AppEndpointOutput::NodeJs { rsc_chunk, .. } => WrittenEndpoint::NodeJs {