--- a/crates/next-api/src/app.rs
+++ b/crates/next-api/src/app.rs
@@ -1636,21 +1636,32 @@ impl Endpoint for AppEndpoint {
                 .project()
                 .emit_all_output_assets(Vc::cell(output_assets));
 
-            let node_root = this.app_project.project().node_root();
-            let server_paths = all_server_paths(output_assets, node_root)
+            let (server_paths, client_paths) = if this
+                .app_project
+                .project()
+                .next_mode()
                 .await?
-                .clone_value();
+                .is_development()
+            {
+                let node_root = this.app_project.project().node_root();
+                let server_paths = all_server_paths(output_assets, node_root)
+                    .await?
+                    .clone_value();
 
-            let client_relative_root = this.app_project.project().client_relative_path();
-            let client_paths = async {
-                anyhow::Ok(
-                    all_paths_in_root(output_assets, client_relative_root)
-                        .await?
-                        .clone_value(),
-                )
-            }
-            .instrument(tracing::info_span!("client_paths"))
-            .await?;
+                let client_relative_root = this.app_project.project().client_relative_path();
+                let client_paths = async {
+                    anyhow::Ok(
+                        all_paths_in_root(output_assets, client_relative_root)
+                            .await?
+                            .clone_value(),
+                    )
+                }
+                .instrument(tracing::info_span!("client_paths"))
+                .await?;
+                (server_paths, client_paths)
+            } else {
+                (vec![], vec![])
+            };
 
             let written_endpoint = match *output {
                 AppEndpointOutput::NodeJs { rsc_chunk, .. } => WrittenEndpoint::NodeJs {