--- a/crates/next-api/src/pages.rs
+++ b/crates/next-api/src/pages.rs
@@ -1273,20 +1273,32 @@ impl Endpoint for PageEndpoint {
                 .emit_all_output_assets(Vc::cell(output_assets));
 
             let node_root = this.pages_project.project().node_root();
-            let server_paths = all_server_paths(output_assets, node_root)
-                .await?
-                .clone_value();
 
-            let client_relative_root = this.pages_project.project().client_relative_path();
-            let client_paths = async {
-                anyhow::Ok(
-                    all_paths_in_root(output_assets, client_relative_root)
-                        .await?
-                        .clone_value(),
-                )
-            }
-            .instrument(tracing::info_span!("client_paths"))
-            .await?;
+            let (server_paths, client_paths) = if this
+                .pages_project
+                .project()
+                .next_mode()
+                .await?
+                .is_development()
+            {
+                let server_paths = all_server_paths(output_assets, node_root)
+                    .await?
+                    .clone_value();
+
+                let client_relative_root = this.pages_project.project().client_relative_path();
+                let client_paths = async {
+                    anyhow::Ok(
+                        all_paths_in_root(output_assets, client_relative_root)
+                            .await?
+                            .clone_value(),
+                    )
+                }
+                .instrument(tracing::info_span!("client_paths"))
+                .await?;
+                (server_paths, client_paths)
+            } else {
+                (vec![], vec![])
+            };
 
             let node_root = &node_root.await?;
             let written_endpoint = match *output {