--- a/crates/next-api/src/paths.rs
+++ b/crates/next-api/src/paths.rs
@@ -1,6 +1,7 @@
 use anyhow::Result;
 use next_core::{all_assets_from_entries, next_manifests::AssetBinding};
 use serde::{Deserialize, Serialize};
+use tracing::Instrument;
 use turbo_rcstr::RcStr;
 use turbo_tasks::{trace::TraceRawVcs, ResolvedVc, TryFlatJoinIterExt, Vc};
 use turbo_tasks_fs::FileSystemPath;
@@ -29,30 +30,35 @@ pub async fn all_server_paths(
     assets: Vc<OutputAssets>,
     node_root: Vc<FileSystemPath>,
 ) -> Result<Vc<ServerPaths>> {
-    let all_assets = all_assets_from_entries(assets).await?;
-    let node_root = &node_root.await?;
-    Ok(Vc::cell(
-        all_assets
-            .iter()
-            .map(|&asset| async move {
-                Ok(
-                    if let Some(path) = node_root.get_path_to(&*asset.ident().path().await?) {
-                        let content_hash = match *asset.content().await? {
-                            AssetContent::File(file) => *file.hash().await?,
-                            AssetContent::Redirect { .. } => 0,
-                        };
-                        Some(ServerPath {
-                            path: path.to_string(),
-                            content_hash,
-                        })
-                    } else {
-                        None
-                    },
-                )
-            })
-            .try_flat_join()
-            .await?,
-    ))
+    let span = tracing::info_span!("all_server_paths");
+    async move {
+        let all_assets = all_assets_from_entries(assets).await?;
+        let node_root = &node_root.await?;
+        Ok(Vc::cell(
+            all_assets
+                .iter()
+                .map(|&asset| async move {
+                    Ok(
+                        if let Some(path) = node_root.get_path_to(&*asset.ident().path().await?) {
+                            let content_hash = match *asset.content().await? {
+                                AssetContent::File(file) => *file.hash().await?,
+                                AssetContent::Redirect { .. } => 0,
+                            };
+                            Some(ServerPath {
+                                path: path.to_string(),
+                                content_hash,
+                            })
+                        } else {
+                            None
+                        },
+                    )
+                })
+                .try_flat_join()
+                .await?,
+        ))
+    }
+    .instrument(span)
+    .await
 }
 
 /// Return a list of relative paths to `root` for all output assets references