Depends on https://github.com/vercel/next.js/pull/73036

Testing using vercel-marketing:
```
before:
    416.31s user 54.20s system 700% cpu 1:07.13 total    
    423.41s user 55.50s system 762% cpu 1:02.80 total
    397.76s user 56.07s system 856% cpu 52.968 total
    410.00s user 58.71s system 796% cpu 58.852 total
after:
    386.88s user 53.57s system 824% cpu 53.410 total
    390.92s user 61.66s system 724% cpu 1:02.48 total
    395.85s user 62.64s system 656% cpu 1:09.84
    389.90s user 52.85s system 802% cpu 55.164 total
```
So about 2.5% faster in CPU time