@ijjk 2024-11-21
## Failing test suites <!--31170750543-->
Commit: 99dddb0c7610f57342f46db42553a2e0215ba6b6

`pnpm test-start-turbo test/e2e/app-dir/actions-allowed-origins/app-action-allowed-origins.test.ts` (turbopack)
- app-dir action allowed origins > should pass if localhost is set as a safe origin

<details>
<summary>Expand output</summary>

  ● app-dir action allowed origins › should pass if localhost is set as a safe origin

    TIMED OUT: hi



    undefined

      718 |
      719 |   if (hardError) {
    > 720 |     throw new Error('TIMED OUT: ' + regex + '\n\n' + content + '\n\n' + lastErr)
          |           ^
      721 |   }
      722 |   return false
      723 | }

      at check (lib/next-test-utils.ts:720:11)
      at Object.<anonymous> (e2e/app-dir/actions-allowed-origins/app-action-allowed-origins.test.ts:25:5)

</details>


Read more about building and testing Next.js in [contributing.md](https://github.com/vercel/next.js/blob/canary/contributing.md).