@mischnic 2024-12-03
Superseded by https://github.com/vercel/next.js/pull/73260